+++ at_secondary/at_persistence_secondary_server/lib/src/notification/at_notification_keystore.dart
@@ -63,27 +63,31 @@ class AtNotificationKeystore
   }
 
   @override
-  Future put(key, value,
+  Future<dynamic> put(key, value,
       {int? time_to_live,
       int? time_to_born,
       int? time_to_refresh,
       bool? isCascade,
       bool? isBinary,
       bool? isEncrypted,
-      String? dataSignature}) async {
+        String? dataSignature,
+        String? sharedKeyEncrypted,
+        String? publicKeyChecksum}) async {
     await _getBox().put(key, value);
     AtNotificationCallback.getInstance().invokeCallbacks(value);
   }
 
   @override
-  Future create(key, value,
+  Future<dynamic> create(key, value,
       {int? time_to_live,
       int? time_to_born,
       int? time_to_refresh,
       bool? isCascade,
       bool? isBinary,
       bool? isEncrypted,
-      String? dataSignature}) async {
+        String? dataSignature,
+        String? sharedKeyEncrypted,
+        String? publicKeyChecksum}) async {
     throw UnimplementedError();
   }